Move compileTalks script in llms-txt.js
