--- docs/scripts/generate-llms/compileTalks.js
@@ -1,29 +1,0 @@
-import fs from 'node:fs';
-import path from 'node:path';
-import ts from 'typescript';
-
-const projectRoot = path.resolve(import.meta.dirname, '../..');
-
-const inputFile = path.join(projectRoot, 'public/static/talks.ts');
-const outputFile = path.join(projectRoot, 'scripts/generate-llms/talks.js');
-
-export async function compileTalksFile() {
-  try {
-    const inputFileContent = fs.readFileSync(inputFile).toString();
-    const outputFileContent = ts.transpileModule(inputFileContent, {
-      module: ts.ModuleKind.ESNext,
-      compilerOptions: {
-        target: 'ES2024',
-        module: 'ES2024',
-        moduleResolution: 'node',
-      },
-    }).outputText;
-
-    fs.writeFileSync(outputFile, outputFileContent, 'utf8');
-
-    console.log(` \x1b[1m\x1b[32m✓\x1b[0m Successfully compiled talks.ts to talks.js`);
-  } catch (error) {
-    console.error('Error compiling talks.ts:', error);
-    process.exit(1);
-  }
-}
