+++ docs/scripts/generate-llms/index.js
@@ -1,12 +1,9 @@
 import process from 'node:process';
 
-import { compileTalksFile } from './compileTalks.js';
 import { generateLlmsEasTxt } from './llms-eas-txt.js';
 import { generateLlmsFullTxt } from './llms-full-txt.js';
 import { generateLlmsSdkTxt } from './llms-sdk.js';
 import { generateLlmsTxt } from './llms-txt.js';
-
-await compileTalksFile();
 
 const results = await Promise.allSettled([
   generateLlmsSdkTxt(),
